UD skimming developments: fix typos, naming, rework UPC candidate analyzer
