+++ PWGUD/TableProducer/CMakeLists.txt
@@ -14,12 +14,12 @@ o2physics_add_dpl_workflow(dgcand-producer
                     PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::AnalysisCore O2Physics::DGCutparHolder
                     COMPONENT_NAME Analysis)
 
-o2physics_add_dpl_workflow(track-skimmer
-        SOURCES UDTrackSkimmer.cxx
-        PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::AnalysisCore
-        COMPONENT_NAME Analysis)
+o2physics_add_dpl_workflow(upctrack-skimmer
+                           SOURCES UPCTrackSkimmer.cxx
+                           PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::AnalysisCore
+                           COMPONENT_NAME Analysis)
 
-o2physics_add_dpl_workflow(candidate-creator
-        SOURCES UDCandidateCreator.cxx
-        PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::AnalysisCore
-        COMPONENT_NAME Analysis)
+o2physics_add_dpl_workflow(upccand-producer
+                           SOURCES UPCandidateProducer.cxx
+                           PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::AnalysisCore
+                           COMPONENT_NAME Analysis)
