+++ PWGUD/Tasks/CMakeLists.txt
@@ -39,12 +39,12 @@ o2physics_add_dpl_workflow(dgcand-analyzer
                            PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::AnalysisCore O2Physics::DGCutparHolder O2Physics::DGPIDSelector
                            COMPONENT_NAME Analysis)
 
-o2physics_add_dpl_workflow(dimuons
-        SOURCES dimuonAnalysis.cxx
-        PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::AnalysisCore O2::DetectorsBase
-        COMPONENT_NAME Analysis)
+o2physics_add_dpl_workflow(upccand-analyzer
+                           SOURCES UPCCandidateAnalyzer.cxx
+                           PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::AnalysisCore O2::DetectorsBase
+                           COMPONENT_NAME Analysis)
 
-o2physics_add_dpl_workflow(track-skimmer-qa
-        SOURCES trackSkimmerQA.cxx
-        PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::AnalysisCore O2::DetectorsBase
-        COMPONENT_NAME Analysis)
+o2physics_add_dpl_workflow(upc-track-skimmer-qa
+                           SOURCES UPCTrackSkimmerQA.cxx
+                           PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::AnalysisCore O2::DetectorsBase
+                           COMPONENT_NAME Analysis)
